+++ go.mod
@@ -3,7 +3,7 @@ module github.com/anexia-it/go-anxcloud
 go 1.15
 
 require (
-	github.com/onsi/ginkgo v1.12.1
+	github.com/onsi/ginkgo v1.16.4
 	github.com/onsi/gomega v1.10.4
 	github.com/stretchr/testify v1.7.0
 	golang.org/x/crypto v0.0.0-20201002170205-7f63de1d35b0
